# PedelecServerSampleSolution/main.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2023-05-19T12:33:48+00:00

from __future__ import annotations

import json
from typing import List, Union

from bson import ObjectId
from fastapi import FastAPI

from models import Pedelec, PedelecFullData
from pymongo.mongo_client import MongoClient

app = FastAPI(
    title='Pedelec API',
    description='CRUD operations for Pedelec',
    version='0.0.1',
    servers=[{'url': 'https://b91c-2a09-80c0-192-0-93d0-44a7-d7c1-b20.ngrok-free.app'}],
)

# Of course, you would usually set the password as environment variable and not directly in the code
uri = "mongodb+srv://<TODO>:<TODO>@<TODO>"

# Create a new client and connect to the server
client = MongoClient(uri)

# get the DB and the collection
pedelec_db = client.get_database("pedelec")
pedelec_collection = pedelec_db.get_collection("pedelec")


@app.get('/pedelecs', response_model=List[PedelecFullData])
def get_pedelecs() -> List[PedelecFullData]:
    """
    Retrieve all Pedelecs
    """
    return [PedelecFullData(**entry) for entry in pedelec_collection.find({})]


@app.post('/pedelecs', response_model=None, responses={'201': {'model': PedelecFullData}})
def post_pedelecs(pedelec: Pedelec) -> Union[None, PedelecFullData]:
    """
    Create a new Pedelec
    """
    pedelec_dict = json.loads(pedelec.json())
    # the id is added to pedelec_dict on insertion in the DB
    pedelec_collection.insert_one(pedelec_dict)
    return PedelecFullData(**pedelec_dict)


@app.get('/pedelecs/{id}', response_model=PedelecFullData)
def get_pedelecs_id(id: str) -> PedelecFullData:
    """
    Get a specific Pedelec by ID
    """
    result = pedelec_collection.find_one({"_id": ObjectId(id)})
    return PedelecFullData(**result)