# openapi.yaml
openapi: 3.0.2
info:
  title: Pedelec API
  description: CRUD operations for Pedelec
  version: 0.0.1
servers:
  - url: http://localhost:8000
paths:
  /pedelecs:
    get:
      summary: Retrieve all Pedelecs
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PedelecFullData'
    post:
      summary: Create a new Pedelec
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pedelec'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedelecFullData'
    
  /pedelecs/{id}:
    get:
      summary: Get a specific Pedelec by ID
      parameters:
        - required: true
          schema:
            type: string
          name: id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedelecFullData'
                  
components:
  schemas:
    Location:
      title: Location
      required:
        - latitude
        - longitude
        - altitude
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
        altitude:
          type: number
    Pedelec:
      title: Pedelec
      required:
        - isAvailable
      type: object
      properties:
        charge:
          title: Charge
          maximum: 100
          minimum: 0
          type: integer
          description: Current charge level of the pedelec, expressed as percentage in (0,100)
        isAvailable:
          type: boolean
        location:
          $ref: '#/components/schemas/Location'
    PedelecFullData:
      allOf:
        - $ref: '#/components/schemas/Pedelec'
        - type: object
          required:
            - id
          properties:
            id:
              type: string
              description: The ID of the pedelec
      description: Full data of the pedelec.